+++ .coveragerc
@@ -577,14 +577,11 @@ omit =
     homeassistant/components/nello/lock.py
     homeassistant/components/nest/__init__.py
     homeassistant/components/nest/api.py
-    homeassistant/components/nest/binary_sensor.py
-    homeassistant/components/nest/camera.py
-    homeassistant/components/nest/camera_legacy.py
-    homeassistant/components/nest/climate.py
-    homeassistant/components/nest/climate_legacy.py
-    homeassistant/components/nest/local_auth.py
-    homeassistant/components/nest/sensor.py
-    homeassistant/components/nest/sensor_legacy.py
+#    homeassistant/components/nest/binary_sensor.py
+#    homeassistant/components/nest/camera.py
+#    homeassistant/components/nest/climate.py
+#    homeassistant/components/nest/sensor.py
+    homeassistant/components/nest/legacy/*
     homeassistant/components/netatmo/__init__.py
     homeassistant/components/netatmo/api.py
     homeassistant/components/netatmo/camera.py
