+++ homeassistant/components/nest/camera.py
@@ -1,9 +1,9 @@
 """Support for Nest cameras that dispatches between API versions."""
 
+from homeassistant.components.nest.legacy.camera import async_setup_legacy_entry
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.typing import HomeAssistantType
 
-from .camera_legacy import async_setup_legacy_entry
 from .camera_sdm import async_setup_sdm_entry
 from .const import DATA_SDM
