+++ homeassistant/components/nest/climate.py
@@ -1,9 +1,9 @@
 """Support for Nest climate that dispatches between API versions."""
 
+from homeassistant.components.nest.legacy.climate import async_setup_legacy_entry
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.typing import HomeAssistantType
 
-from .climate_legacy import async_setup_legacy_entry
 from .climate_sdm import async_setup_sdm_entry
 from .const import DATA_SDM
