+++ homeassistant/components/nest/const.py
@@ -1,6 +1,7 @@
 """Constants used by the Nest component."""
 
 DOMAIN = "nest"
+DATA_NEST = "nest"
 DATA_SDM = "sdm"
 DATA_SUBSCRIBER = "subscriber"
