+++ homeassistant/components/nest/legacy/const.py
@@ -1,0 +1,3 @@
+"""Constants used by the legacy Nest component."""
+
+SIGNAL_NEST_UPDATE = "nest_update"
