+++ homeassistant/components/nest/legacy/local_auth.py
@@ -4,11 +4,10 @@ from functools import partial
 
 from nest.nest import AUTHORIZE_URL, AuthorizationError, NestAuth
 
+from homeassistant.components.nest import config_flow
+from homeassistant.components.nest.const import DOMAIN
 from homeassistant.const import HTTP_UNAUTHORIZED
 from homeassistant.core import callback
-
-from . import config_flow
-from .const import DOMAIN
 
 
 @callback
