+++ homeassistant/components/nest/legacy/sensor.py
@@ -1,8 +1,14 @@
 """Support for Nest Thermostat sensors for the legacy API."""
 import logging
 
+from homeassistant.components.nest.legacy import (
+    DATA_NEST,
+    DATA_NEST_CONFIG,
+    NestSensorDevice,
+)
 from homeassistant.const import (
     CONF_MONITORED_CONDITIONS,
+    CONF_SENSORS,
     DEVICE_CLASS_HUMIDITY,
     DEVICE_CLASS_TEMPERATURE,
     PERCENTAGE,
@@ -10,8 +16,6 @@ from homeassistant.const import (
     TEMP_CELSIUS,
     TEMP_FAHRENHEIT,
 )
-
-from . import CONF_SENSORS, DATA_NEST, DATA_NEST_CONFIG, NestSensorDevice
 
 SENSOR_TYPES = ["humidity", "operation_mode", "hvac_state"]
