+++ homeassistant/components/nest/sensor.py
@@ -1,10 +1,10 @@
 """Support for Nest sensors that dispatches between API versions."""
 
+from homeassistant.components.nest.legacy.sensor import async_setup_legacy_entry
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.typing import HomeAssistantType
 
 from .const import DATA_SDM
-from .sensor_legacy import async_setup_legacy_entry
 from .sensor_sdm import async_setup_sdm_entry
