+++ tests/components/nest/test_local_auth.py
@@ -4,7 +4,8 @@ from urllib.parse import parse_qsl
 import pytest
 import requests_mock as rmock
 
-from homeassistant.components.nest import config_flow, const, local_auth
+from homeassistant.components.nest import config_flow, const
+from homeassistant.components.nest.legacy import local_auth
 
 
 @pytest.fixture
